Fetches real-time temperature from API
Updates the application to fetch real-time temperature data from an external API, replacing the simulated temperature.

This improves the accuracy and realism of the oven simulation.
Removes unused atom related to current temperature.

Also sets a fixed max ingredient temp to 100, and updates the temp refetch to every 1000ms for more frequent updates.